refactor: split transformStreams file by extracting utils in separate file
