+++ transformStreams.js
@@ -1,40 +1,5 @@
 const { Transform } = require('stream');
-
-const helper = (str, step) => {
-  return [...str].reduce((acc, item) => {
-    const pos = item.charCodeAt();
-    if (item.match(/[A-Z]/)) {
-      return (acc += String.fromCharCode(((pos - 65 + 26 + step) % 26) + 65));
-    } else if (item.match(/[a-z]/)) {
-      return (acc += String.fromCharCode(((pos - 97 + 26 + step) % 26) + 97));
-    }
-
-    return (acc += item);
-  }, '');
-};
-
-const caesar = (str, code) => {
-  const step = code ? 1 : -1;
-  return helper(str, step);
-};
-
-const rot8 = (str, code) => {
-  const step = code ? 8 : -8;
-  return helper(str, step);
-};
-
-const atbash = (str) => {
-  return [...str].reduce((acc, item) => {
-    const pos = item.charCodeAt();
-    if (item.match(/[A-Z]/)) {
-      return (acc += String.fromCharCode(Math.abs(pos - 65 - 26) + 64));
-    } else if (item.match(/[a-z]/)) {
-      return (acc += String.fromCharCode(Math.abs(pos - 97 - 26) + 96));
-    }
-
-    return (acc += item);
-  }, '');
-};
+const utils = require('./utils');
 
 class C1 extends Transform {
   constructor() {
@@ -42,7 +7,7 @@ class C1 extends Transform {
   }
 
   _transform(chunk, encoding, callback) {
-    const transformChunk = caesar(chunk.toString('utf8'), 1);
+    const transformChunk = utils.caesar(chunk.toString('utf8'), 1);
     this.push(transformChunk);
     callback();
   }
@@ -54,7 +19,7 @@ class C0 extends Transform {
   }
 
   _transform(chunk, encoding, callback) {
-    const transformChunk = caesar(chunk.toString('utf8'), 0);
+    const transformChunk = utils.caesar(chunk.toString('utf8'), 0);
     this.push(transformChunk);
     callback();
   }
@@ -66,7 +31,7 @@ class R1 extends Transform {
   }
 
   _transform(chunk, encoding, callback) {
-    const transformChunk = rot8(chunk.toString('utf8'), 1);
+    const transformChunk = utils.rot8(chunk.toString('utf8'), 1);
     this.push(transformChunk);
     callback();
   }
@@ -78,7 +43,7 @@ class R0 extends Transform {
   }
 
   _transform(chunk, encoding, callback) {
-    const transformChunk = rot8(chunk.toString('utf8'), 0);
+    const transformChunk = utils.rot8(chunk.toString('utf8'), 0);
     this.push(transformChunk);
     callback();
   }
@@ -90,7 +55,7 @@ class A extends Transform {
   }
 
   _transform(chunk, encoding, callback) {
-    const transformChunk = atbash(chunk.toString('utf8'));
+    const transformChunk = utils.atbash(chunk.toString('utf8'));
     this.push(transformChunk);
     callback();
   }
